% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcvExports.R
\name{dDeta}
\alias{dDeta}
\title{Obtaining derivative w.r.t. linear predictor}
\usage{
dDeta(y, mu, wt, theta, fam, deriv = 0)
}
\arguments{
\item{y}{vector of observations.}

\item{mu}{if \code{eta} is the linear predictor, \code{mu = inv_link(eta)}. In a traditional GAM \code{mu=E(y)}.}

\item{wt}{vector of weights.}

\item{theta}{vector of family parameters that are not regression coefficients (e.g. scale parameters).}

\item{fam}{the family object.}

\item{deriv}{the order of derivative of the smoothing parameter score required.}
}
\value{
A list of derivatives.
}
\description{
INTERNAL function. Distribution families provide derivatives of the deviance and link w.r.t. \code{mu = inv_link(eta)}.
             This routine converts these to the required derivatives of the deviance w.r.t. eta, the linear predictor.
}
\author{
Simon N. Wood <simon.wood@r-project.org>.
}
