\name{get.family}
\alias{get.family}
\title{Identifies families}
\description{
  This is an internal function of package \code{mgcv} used to determine which family is intended within a call to \code{gam()}.}
\author{ Mike Lonergan\email{mel@st-and.ac.uk} }

\seealso{
  \code{\link{family}} .
}

\keyword{models}
\keyword{regression}
\keyword{smooth}