\name{print.gam}
\alias{print.gam}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Generalized Additive Model default print statement}
\description{ This is the default print statement for a GAM object. If you need a list of everything
that is part of a gam object see \code{\link{gam}}, or use \code{names()}. The family (including link), 
model formula, and estimated degrees of freedom for each model term (plus total) are printed, as 
well as the minimized GCV or UBRE/AIC score, depending on which was used.
}


\references{

Wood, S.N. (2006) Generalized Additive Models: An Introduction with R. CRC/
Chapmand and Hall, Boca Raton, Florida.

\url{http://www.maths.bath.ac.uk/~sw283/}
}
\author{ Simon N. Wood \email{simon.wood@r-project.org} }



\seealso{   \code{\link{gam}},\code{\link{mgcv}}}

\keyword{models} \keyword{smooth} \keyword{regression}%-- one or more ..









